Extract the agents codec into cli/integrations/
Move the agent (subagent) artifact codec out of cli/commands/setup.py
into a dedicated cli/integrations/agents.py, continuing the setup
subsystem decomposition. _claude_agent_dir and materialize_agents move
verbatim alongside their explanatory comment, and setup.py imports
materialize_agents from the new module. The write-safety test retargets
its materialize_agents import to the new location.

This is a pure verbatim move with zero behavior change. The function
bodies are byte-identical to their prior form, including the lazy
nauro.agents import that preserves the existing import-cycle behavior
and the deliberately non-atomic write path for agent files.
